undo Super Patcher Mavic Pro v01.04.0300
This includes a stock 305 module, a stock 306, and stock cfg.sig for Mavic Pro firmware version 01.04.0300. Since these are the modules Super-Patcher deals with you can quickly flash this firmware between custom flight controller modules like Super Patcher so that will bring the firmware back to a stock v01.04.0300 so that Super-Patcher can be reflashed or another custom flight controller can be flashed to Mavic. 

This is to save time. If there's and issue you are trying to solve you can try this FW version but it isnt as good as flashing a full version of stock 01.04.0300. Its just a time saver